refactor: DiaryComponent 제거 후 Diary 페이지로 이동
