+++ src/components/pages/diary/diary/modal/index.ts
@@ -1,2 +1,2 @@
-export { default as Modal } from "./Modal";
+export { default as ModalContainer } from "./ModalContainer";
 export * from "./element";
